Added UIManager script, doesn't completely work yet but it ain't broken.
Had to mess with other scripts to make it work. Hope I didn't fuck anything up!
Fixed checkpoints with tags and colliders. Also added a panel which enables the player to enter MM (mutationmanager) without getting an orb. The panel is supposed to be set to active when Gus is inside the collider of a checkpoint, and I suppose set to NOT active when he's not. Right now it's set to active but won't go away.
We also have a new, shiny scene. Check it out.