Finished filling out sides of level, and added MUSIC to the scenes!
Added new block 1x3 (with an own material with different tiling) to make it easier. Fixed the pictures on MMmanager. The background sprite seem to be transparent. Y THO?? :thinking:
Need to fix: Animation on credits-roll need to be longer since the text is longer now. Also a bit slower. Add the terminals, and the orbs where they're supposed to be. WIN-flag as well.